+++ EK_Component/DataStruct/List/EK_List.h
@@ -30,7 +30,7 @@ typedef struct EK_Node_t
     struct EK_Node_t *Node_Prev; /**< 前一个节点 */
     struct EK_Node_t *Node_Next; /**< 后一个节点 */
     EK_List_t *Node_Owner; /**< 节点所有者 */
-    uint32_t Node_Order; /**< 节点序号 */
+    uint16_t Node_Order; /**< 节点序号 */
     bool Node_isDynamic; /**< 是否来自动态分配 */
 } EK_Node_t;
 
@@ -42,8 +42,8 @@ typedef struct EK_List_t
 } EK_List_t;
 
 /* ========================= 函数声明区 ========================= */
-EK_Result_t EK_rNodeCreate_Static(EK_Node_t *node, void *content, uint32_t order);
-EK_Node_t *EK_pNodeCreate_Dynamic(void *content, uint32_t order);
+EK_Result_t EK_rNodeCreate_Static(EK_Node_t *node, void *content, uint16_t order);
+EK_Node_t *EK_pNodeCreate_Dynamic(void *content, uint16_t order);
 EK_Result_t EK_rListCreate_Static(EK_List_t *list, EK_Node_t *dummy_node);
 EK_List_t *EK_pListCreate_Dynamic(void);
 EK_Result_t EK_rListInsertEnd(EK_List_t *list, EK_Node_t *node);
